+++ src/Filesystem.php
@@ -280,6 +280,18 @@ class Filesystem extends BaseFilesystem
     }
 
     /**
+     * Given an existing path, convert it to a path relative to a given starting path
+     * @param string $endPath Absolute path of target
+     * @param string $startPath Absolute path where traversal begins
+     * @return string Path of target relative to starting path
+     * @since 1.5.10
+     */
+    public function makePathRelative(string $endPath, string $startPath): string
+    {
+        return rtrim(parent::makePathRelative($endPath, $startPath), '/\\');
+    }
+
+    /**
      * Normalizes the path, applying the right slash term
      * @param string $path Path you want normalized
      * @return string Normalized path
@@ -319,9 +331,11 @@ class Filesystem extends BaseFilesystem
      */
     public function rtr(string $path): string
     {
-        $root = $this->getRoot();
-        if ($this->isAbsolutePath($path) && str_starts_with($path, $root)) {
-            $path = $this->normalizePath($this->makePathRelative($path, $root));
+        if ($this->isAbsolutePath($path)) {
+            $root = $this->getRoot();
+            if (str_starts_with($path, $root)) {
+                $path = $this->normalizePath($this->makePathRelative($path, $root));
+            }
         }
 
         return rtrim($path, DS);